refactor(layout): extract heading classification
Move the heading detection heuristics into a dedicated layout module so the layout entrypoint can focus on line grouping and orchestration.
